# src/documentation/swagger-stag.yaml
openapi: 3.0.3
info:
  title: Careplace - Backend REST API
  description: |-
    Careplace REST API Documentation - OpenAPI Specification 3.0.3.
    This REST API is used to manage all Careplace applications.
  termsOfService: "https://www.careplace.pt/termos-e-condicoes"
  contact:
    email: admin@careplace.pt
  version: 1.0.0
servers:
  - url: "http://localhost:8080/api/v1"
    description: Development Environment
    default: true
  - url: "http://www.staging.api.careplace.pt"
    description: Staging Environment
...
paths:
...
  "/companies/orders/{orderId}/send-quote":
    post:
      tags:
        - Companies
      summary: Send quote
      description: Send quote by id.
      operationId: sendQuote
      security:
        - jwt_context:
            - write
            - read
      parameters:
        - name: orderId
          in: path
          description: Order id
          required: true
          schema:
            type: string
          example: 5e1f9b9b9b9b9b9b9b9b9b9b
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    example: 5f7f9f9f9f9f9f9f9f9f9f9f
                  company:
                    type: string
                    example: 5f7f9f9f9f9f9f9f9f9f9f9f
                  user:
                    type: string
                    example: 5f7f9f9f9f9f9f9f9f9f9f9f
                  relative:
                    type: string
                    example: 5f7f9f9f9f9f9f9f9f9f9f9f
                  status:
                    type: string
                    example: payment_new
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"
                  schedule_information:
                    type: object
                    properties:
                      start_date:
                        type: date
                        example: 2020-10-10T00:00:00.000Z
                      end_date:
                        type: date
                        example: null
                      reccurency_type:
                        type: number
                        example: 1
                      schedule:
                        type: array
                        items:
                          type: object
                          properties:
                            week_day:
                              type: number
                              example: 1
                            start:
                              type: date
                              example: 2020-10-10T00:00:00.000Z
                            end:
                              type: date
                              example: 2020-10-10T00:00:00.000Z
                  address:
                    type: object
                    properties:
                      street:
                        type: string
                        example: 123 Main St
                      city:
                        type: string
                        example: New York
                      postal_code:
                        type: string
                        example: 1234-567
                      state:
                        type: string
                        example: NY
                      zip:
                        type: string
                        example: 10001
                      country:
                        type: string
                        example: US
                      coordinates:
                        type: object
                        properties:
                          latitute:
                            type: number
                            example: 38.736946
                          longitude:
                            type: number
                            example: -9.142685
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Missing or invalid token.
                  type:
                    type: string
                    example: INVALID_REQUEST
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Access token is expired.
                  type:
                    type: string
                    example: INVALID_REQUEST
...
  /users/payment-methods:
    get:
      tags:
        - Users
      summary: Get all payment methods
      description: Get all payment methods for the logged in user.
      operationId: getPaymentMethods
      security:
        - jwt_context:
            - write
            - read
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: tok_1F9f9f9f9f9f9f9f9f9f9f9f
                    object:
                      type: string
                      example: token
                    card:
                      type: object
                      properties:
                        object:
                          type: string
                          example: payment_method
                        address_city:
                          type: string
                          example: null
                        address_country:
                          type: string
                          example: null
                        address_line1:
                          type: string
                          example: null
                        address_line1_check:
                          type: string
                          example: null
                        address_line2:
                          type: string
                          example: null
                        address_state:
                          type: string
                          example: null
                        address_zip:
                          type: string
                          example: null
                        address_zip_check:
                          type: string
                          example: null
                        brand:
                          type: string
                          example: Visa
                        country:
                          type: string
                          example: US
                        cvc_check:
                          type: string
                          example: null
                        dynamic_last4:
                          type: string
                          example: null
                        exp_month:
                          type: number
                          example: 12
                        exp_year:
                          type: number
                          example: 2022
                        funding:
                          type: string
                          example: credit
                        last4:
                          type: string
                          example: 4242
                        metadata:
                          type: object
                          properties: {}
                        name:
                          type: string
                          example: null
                        tokenization_method:
                          type: string
                          example: null
                    created:
                      type: integer
                      example: 1591234567
                    livemode:
                      type: boolean
                      example: false
                    type:
                      type: string
                      example: card
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Missing or invalid token.
                  type:
                    type: string
                    example: INVALID_REQUEST
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Access token is expired.
                  type:
                    type: string
                    example: INVALID_REQUEST
    post:
      tags:
        - Users
      summary: Create a payment method
      description: >-
        Create a payment method. At the moment it only supports credit/debit
        cards. The user to each the payment method is attached to is the user
        extracted from the access token. To create the payment_method_id the
        Client Application needs to use the Stripe JS library. @see
        https://stripe.com/docs/stripe-js
      operationId: createPaymentMethod
      security:
        - jwt_context:
            - write
            - read
      requestBody:
        description: Create a payment method
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_method_id:
                  type: string
                  example: tok_1F9f9f9f9f9f9f9f9f9f9f9f
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: tok_1F9f9f9f9f9f9f9f9f9f9f9f
                  object:
                    type: string
                    example: token
                  billing_details:
                    type: object
                    properties:
                      address:
                        type: object
                        properties:
                          city:
                            type: string
                            example: null
                          country:
                            type: string
                            example: null
                          line1:
                            type: string
                            example: null
                          line2:
                            type: string
                            example: null
                          postal_code:
                            type: string
                            example: null
                          state:
                            type: string
                            example: null
                      email:
                        type: string
                        example: null
                      name:
                        type: string
                        example: null
                      phone:
                        type: string
                        example: null
                  card:
                    type: object
                    properties:
                      object:
                        type: string
                        example: payment_method
                      brand:
                        type: string
                        example: visa
                      checks:
                        type: object
                        properties:
                          address_line1_check:
                            type: string
                            example: null
                          address_postal_code_check:
                            type: string
                            example: null
                          cvc_check:
                            type: string
                            example: null
                      country:
                        type: string
                        example: US
                      exp_month:
                        type: number
                        example: 12
                      exp_year:
                        type: number
                        example: 2022
                      funding:
                        type: string
                        example: credit
                      generated_from:
                        type: string
                        example: null
                      last4:
                        type: string
                        example: 4242
                      networks:
                        type: object
                        properties:
                          available:
                            type: array
                            items:
                              type: string
                              example: visa
                          preferred:
                            type: string
                            example: null
                      three_d_secure_usage:
                        type: object
                        properties:
                          supported:
                            type: boolean
                            example: true
                      wallet:
                        type: string
                        example: null
                  created:
                    type: integer
                    example: 1591234567
                  type:
                    type: string
                    example: card
...
components:
  schemas:
...
    Service:
      type: object
      properties:
        _id:
          type: string
          example: 5e1f9b9b9b9b9b9b9b9b9b9b
        name:
          type: string
          example: Bathing
        description:
          type: string
          example: Bathing
        short_description:
          type: string
          example: Bathing
        image:
          type: string
          example: "https://www.google.com"
  securitySchemes:
    jwt_context:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token